+++ src/main/scala/info/gratour/jt808common/JT808Consts.java
@@ -1,5 +1,7 @@
 package info.gratour.jt808common;
 
+import info.gratour.common.error.ErrorWithCode;
+
 public class JT808Consts {
 
     public static class ProtocolVersions {
@@ -8,6 +10,17 @@ public class JT808Consts {
 
         public static boolean isValidProtoVer(int ver) {
             return ver == PROTO_VER__REV2013 || ver == PROTO_VER__REV2019;
+        }
+
+        public static int detectProtoVerBySimNoLen(int simNoLen) {
+            switch (simNoLen) {
+                case 12:
+                    return PROTO_VER__REV2013;
+                case 20:
+                    return PROTO_VER__REV2019;
+                default:
+                    throw ErrorWithCode.internalError("Unrecognized simNoLen: " + simNoLen);
+            }
         }
     }
 }