+++ doc/change-log.md
@@ -29,6 +29,7 @@
 - 新增 `SetTraceMsg` 类
 - `TermCmd` 增加 `setParamsJson`, `setAckParamsJson` 方法
 - 新增 `QryTermCmdReq` 类
+- 新增 `info.gratour.jt808common.JT808Consts.ProtocolVersions.detectProtoVerBySimNoLen` 方法
 
 ## 不兼容变更
 - `AvUploadReq` 增加 `fileSz` 必填属性